Add delete product view and delete confirmation template

+++ products/views.py
@@ -125,8 +125,16 @@ def delete_product(request, product_id):
     if not request.user.is_superuser:
         messages.error(request, 'Sorry, only store owners can do that.')
         return redirect(reverse('home'))
-        
+
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted successfully')
-    return redirect(reverse('products'))
+    if request.method == 'POST':    
+        product.delete()
+        messages.success(request, 'Product deleted successfully')
+        return redirect(reverse('products'))
+
+    template = 'delete_product.html'
+    context = {
+        "product": product,
+    }
+
+    return render(request, template, context)
